Initial version of authentication code
Currently, I am using test code in `src/main.rs`, but it does allow
authentication and sessions. However, due to issues with Warp, requests
that pass session IDs which are valid UUIDs but not in use get HTTP 500
with an "unhandled rejection", rather than the 401 Unauthorized that
they should be getting.